backend/requirements.txt
flask==3.1.0
twilio==9.4.0
requests==2.32.3
python-dotenv==1.1.0
flask-socketio==5.5.1
flask-sock>=0.7.0
pydub==0.25.1
pyngrok==7.2.2
websocket-client==1.8.0
webrtcvad>=2.0.10
audioop-lts>=0.2.1
